End game goal was added (treasure chest!)
Chest is currently located about 10 tiles to the right of the player's start position

When the player collides with the chest, it displays a Form saying who won and a button to play again which resets the current playing field to its original values for a new game.

In forms.js, I added the ability to just display text by passing "basic-text" as the type.

Bug fix in player.js, unbinding mousemove.